Fix(Structure): File structure fix
