--- public/_redirects
@@ -1,2 +1,0 @@
-# Redirects from what the browser requests to what we serve
-/*    /index.html   200
