+++ src/App.tsx
@@ -1,8 +1,7 @@
 import { Routes, Route } from "react-router-dom";
 import Home from "./pages/Home/Home";
-import Login from "./pages/Login/Login";
-import Sponsor from "./pages/Sponsor/sponsor";
-import Team from "./pages/Team/team";
+import Sponsor from "./pages/Sponsor/Sponsor";
+import Team from "./pages/Team/Team";
 
 
 const App: React.FC = () => {
@@ -11,7 +10,6 @@ const App: React.FC = () => {
       <Route path="/" element={<Home />} />
       <Route path="/home" element={<Home />} />
       <Route path="/team" element={<Team />} />
-      <Route path="/login" element={<Login />} />
       <Route path="/sponsor" element={<Sponsor />} />
     </Routes>
   );
