--- src/components/Why.tsx
@@ -1,10 +1,0 @@
-
-const Why = () => {
-  return (
-    <div className="relative">
-      
-    </div>
-  );
-};
-
-export default Why;
