--- src/pages/Login/Login.tsx
@@ -1,9 +1,0 @@
-const Login: React.FC = () => {
-  return (
-    <div>
-      <h1>Login Page</h1>
-    </div>
-  );
-};
-
-export default Login;
