+++ src/pages/Sponsor/sponsor.css
@@ -1,0 +1,42 @@
+@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;700&display=swap');
+
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+@keyframes fadeIn {
+  from { opacity: 0.2; }
+  to { opacity: 1; }
+}
+
+body {
+  font-family: 'Poppins', sans-serif;
+  background-color: #FFF; /* Fallback color */
+  
+}
+
+
+
+/* SponsorPage.css */
+
+
+::-webkit-scrollbar {
+  width: 5px;
+  height: 12px; 
+}
+
+::-webkit-scrollbar-track {
+  background: black;
+}
+
+::-webkit-scrollbar-thumb {
+  background: linear-gradient(45deg, #065f46, #4ade80);
+}
+
+::-webkit-scrollbar-thumb:hover {
+  background: linear-gradient(45deg, #067152, #52ec8b);
+}
+
+::-webkit-scrollbar-corner {
+  background: black;
+}
