+++ src/style.css
@@ -25,7 +25,7 @@
   left: 0;
   width: 100%;
   height: 100%;
-  background-image: url('/src/assets/neo2.jpg'); /* Adjust path as needed */
+  background-image: url('/src/assets/Neo-bg.jpg');
   background-size: cover;
   background-repeat: no-repeat;
   background-position: center;
